requirements.txt
langchain
requests
streamlit
python-dotenv
beautifulsoup4
nltk
sentence_transformers
faiss-cpu